+++ helpers/deployments/types.ts
@@ -1,7 +1,3 @@
-interface ChainList<VaultType> {
-    [vault: string]: VaultType;
-}
-
 export interface ContractInitList<VaultType> {
-    [chain: string]: ChainList<VaultType>;
+    [chain: string]: VaultType;
 }
