+++ helpers/deployments/vaults/VaultUniswapV2/deployment.ts
@@ -1,0 +1,42 @@
+import { chains, vaultFees, zeroAddress } from "../../../constants";
+import { ContractInitList } from "../../types";
+import { VaultUniswapV2Init } from "./types";
+
+const contractInits: ContractInitList<VaultUniswapV2Init> = {
+    matic: {
+        priceFeeds: {
+            eth: chains.matic!.priceFeeds.matic,
+            stablecoin: chains.matic!.priceFeeds.usdc,
+        },
+        baseInit: {
+            treasury: chains.matic!.admin.multiSigOwner,
+            router: chains.matic!.infra.uniRouterAddress,
+            stablecoin: chains.matic!.tokens.usdc,
+            tokenWETH: chains.matic!.tokens.wmatic,
+            entranceFeeFactor: vaultFees.entranceFeeFactor,
+            withdrawFeeFactor: vaultFees.withdrawFeeFactor,
+        },
+    },
+    avalanche: {
+        priceFeeds: {
+            eth: chains.avalanche!.priceFeeds.avax,
+            stablecoin: chains.avalanche!.priceFeeds.usdc,
+        },
+        baseInit: {
+            treasury: chains.avalanche!.admin.multiSigOwner,
+            router: chains.avalanche!.infra.uniRouterAddress,
+            stablecoin: chains.avalanche!.tokens.usdc,
+            tokenWETH: chains.avalanche!.tokens.wavax,
+            entranceFeeFactor: vaultFees.entranceFeeFactor,
+            withdrawFeeFactor: vaultFees.withdrawFeeFactor,
+        },
+    }
+};
+
+export const deploymentArgs = (chain: string, timelockOwner: string, gov: string) => {
+    return [
+        contractInits[chain],
+        timelockOwner,
+        gov
+    ];
+};
